refactor: 메뉴와 메뉴 그룹 패키지 분리
