+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -8,8 +8,8 @@ import static org.junit.jupiter.params.provider.EnumSource.Mode.EXCLUDE;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.menu.MenuGroup;
-import kitchenpos.domain.menu.MenuGroupRepository;
+import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
 import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.order.Order;
